Update __init__.py
only ask for resume if it first attempt

+++ jesse_optuna/__init__.py
@@ -79,35 +79,53 @@ def run() -> None:
     # Initialize retry parameters
     retry_delay = 1  # Start with a 1 second delay
     max_delay = 600  # Maximum delay of 10 minutes
+    first_attempt = True
 
     while True:
         try:
-            try:
-                study = JoblibStudy(study_name=study_name, direction="maximize", sampler=sampler,
-                                    storage=storage, load_if_exists=False)
-            except optuna.exceptions.DuplicatedStudyError:
-                if click.confirm('Previous study detected. Do you want to resume?', default=True):
-                    study = JoblibStudy(study_name=study_name, direction="maximize", sampler=sampler,
-                                        storage=storage, load_if_exists=True)
-                elif click.confirm('Delete previous study and start new?', default=False):
-                    optuna.delete_study(study_name=study_name, storage=storage)
+            if first_attempt:
+                try:
                     study = JoblibStudy(study_name=study_name, direction="maximize", sampler=sampler,
                                         storage=storage, load_if_exists=False)
-                else:
-                    print("Exiting.")
-                    return
-            
-            # Set study user attributes
-            study.set_user_attr("strategy_name", cfg['strategy_name'])
-            study.set_user_attr("exchange", cfg['exchange'])
-            study.set_user_attr("symbol", cfg['symbol'])
-            study.set_user_attr("timeframe", cfg['timeframe'])
-
-            # Start optimization
-            study.optimize(objective, n_jobs=cfg['n_jobs'], n_trials=cfg['n_trials'])
-
-            # If optimization is successful, break out of the loop
-            break
+                except optuna.exceptions.DuplicatedStudyError:
+                    if click.confirm('Previous study detected. Do you want to resume?', default=True):
+                        study = JoblibStudy(study_name=study_name, direction="maximize", sampler=sampler,
+                                            storage=storage, load_if_exists=True)
+                    elif click.confirm('Delete previous study and start new?', default=False):
+                        optuna.delete_study(study_name=study_name, storage=storage)
+                        study = JoblibStudy(study_name=study_name, direction="maximize", sampler=sampler,
+                                            storage=storage, load_if_exists=False)
+                    else:
+                        print("Exiting.")
+                        return
+                
+                # Set study user attributes
+                study.set_user_attr("strategy_name", cfg['strategy_name'])
+                study.set_user_attr("exchange", cfg['exchange'])
+                study.set_user_attr("symbol", cfg['symbol'])
+                study.set_user_attr("timeframe", cfg['timeframe'])
+    
+                # Start optimization
+                study.optimize(objective, n_jobs=cfg['n_jobs'], n_trials=cfg['n_trials'])
+                first_attempt = False
+    
+                # If optimization is successful, break out of the loop
+                break
+            else:
+                study = JoblibStudy(study_name=study_name, direction="maximize", sampler=sampler,
+                                            storage=storage, load_if_exists=True)
+                # Set study user attributes
+                study.set_user_attr("strategy_name", cfg['strategy_name'])
+                study.set_user_attr("exchange", cfg['exchange'])
+                study.set_user_attr("symbol", cfg['symbol'])
+                study.set_user_attr("timeframe", cfg['timeframe'])
+    
+                # Start optimization
+                study.optimize(objective, n_jobs=cfg['n_jobs'], n_trials=cfg['n_trials'])
+                first_attempt = False
+    
+                # If optimization is successful, break out of the loop
+                break
         except OperationalError as e:
             if 'the database system is shutting down' in str(e) or 'server closed the connection unexpectedly' in str(e):
                 # Handle expected database disconnection errors.
